Defend API Completed

+++ core/src/com/ninja/game/Sprite/ICharater.java
@@ -4,13 +4,13 @@ package com.ninja.game.Sprite;
  * Created by ather on 19/11/2559.
  */
 public interface ICharater {
-    public int getHealth();
-    public int getMana();
-    public int getStack();
+    public double getHealth();
+    public double getMana();
+    public double getStack();
     public double getDamage();
 
     //Action
-    void attack(Character character, int dmg);
-    void attacked(int dmg);
-    void heal(int hpStack);
+    void attack(Character character, double dmg);
+    void attacked(double dmg);
+    void heal(double hpStack);
 }
